+++ CHANGELOG.md
@@ -147,6 +147,7 @@
 * nostr: deprecate `Keys::vanity` ([Yuki Kishimoto])
 * database: deprecate `NostrEventsDatabase::event_id_seen` ([Yuki Kishimoto])
 * database: deprecate `NostrEventsDatabase::event_seen_on_relays` ([Yuki Kishimoto])
+* sdk: deprecate `Options::req_filters_chunk_size` ([Yuki Kishimoto])
 
 ## [v0.38.0] - 2024/12/31